input: cs40l26: Use LAST_WAKESRC_CTL to determine GPIO status
GPIO information is lost when the device enters hibernation.
Because of this, erroneous interrupts are issued upon wakeup in
numerous situations.

The driver will now use the LAST_WAKESRC_CTL firmware field,
which reports the true status of the GPIOs prior to
entering hibernate mode, to identify incorrect or missed GPIO
trigger interrupts and report accurate information to the user.

Change-Id: I3197dc8823c09b9b6b4723e3eaacc42f42e6c58a
Signed-off-by: Fred Treven <fred.treven@cirrus.com>