input: cs40l26: Changes to PM hibernate/shutdown timeout
Change the default timeout value for the
standby->hibernate and standby->shutdown transitions to
100 ms from 5000 ms such that the ALLOW_HIBERNATE
mailbox command more closely corresponds to when the
device enters hibernate. Lowering this value to less than
100 ms risks entering hibernate too quickly while
the driver is still carrying out some process such as
verifying the mailbox command was acknowledged.

It is possible that this minimum timeout value is still
too short on some systems. For those cases, a read/write
sysfs control is made available to change the timeout to
an acceptable value.

The timeout duration was being changed in the dsp_shutdown
function, but has been removed because a) the default value
will now be short enough to avoid a special SHUTDOWN case and
b) attempting to acknowledge the SHUTDOWN command after the
timeout will never work because the device will need to wake
up to report this value.

Change-Id: I5277f40cbc73c3f4c86b4dfb45081c971a4ed5ae
Signed-off-by: Fred Treven <fred.treven@cirrus.com>