BACKPORT: media: uapi: move VP8 stateless controls out of staging
Until now, the VP8 V4L2 API was not exported as a public API,
and only defined in a private media header (media/vp8-ctrls.h).

The reason for this was a concern about the API not complete
and ready to support VP8 decoding hardware accelerators.

After reviewing the VP8 specification in detail, and now
that the API is able to support Cedrus and Hantro G1,
we can consider this ready.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>

(cherry picked from commit 363240ce1c08875815d28276f0a793bcaedb1ee9)

Conflicts:
   Documentation/userspace-api/media/v4l/ext-ctrls-codec-stateless.rst
   Documentation/userspace-api/media/v4l/ext-ctrls-codec.rst

BUG=b:187666859
TEST=build for cherry

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Ib6eda74516eb9fc069be5b2159744d5c5cb9d25d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2905330
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>