input: cs40l26: Improvements to pm_state_transition
For the shutdown and allow hibernate cases, there is a risk
that the part switches modes prior to the driver reading back
the acknowledge value from the mailbox. In these instances, the
ack_read would wake the part again causing the driver to mislabel
the device's PM state. Avoid reading back from the mailbox in these
cases and assume that the transition happens succesfully.

Some formatting changes to simplify the code.

Change-Id: Ie522344cc0c52c9e179e978f558a68ab4c92bf26
Signed-off-by: Fred Treven <fred.treven@cirrus.com>