UPSTREAM: PM / devfreq: Register devfreq as a cooling device on demand
Currently the default behavior is to manually having the devfreq
backend to register themselves as a devfreq cooling device.

Instead of adding the code in the drivers for the thermal cooling
device registering, let's provide a flag in the devfreq's profile to
tell the common devfreq code to register the newly created devfreq as
a cooling device.

Suggested-by: Chanwoo Choi <cwchoi00@gmail.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Reviewed-by: Steven Price <steven.price@arm.com>
Reviewed-by: Lukasz Luba <lukasz.luba@arm.com>
Signed-off-by: Chanwoo Choi <cw00.choi@samsung.com>

(cherry picked from commit 1224451bb6f938023dd7fa4e7ba43bfb185bc9e3)

BUG=none
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I5cc7408323d964b2b9057248ad97067965914241
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2874908
Reviewed-by: Sean Paul <seanpaul@chromium.org>