ASoC: cs35l45: Add support for DSP logging buffer count
When raising the MBOX4 IRQ, the firmware sends an increment count
to the driver where its value must be checked by the driver. The
count should not be the same as the previous one and should not be 0.
If the count is intermittent, which indicates there has been buffer
lost, the driver must report this as a warning. Currently, there is
no way the driver could recover the lost buffer.

Change-Id: I9f38cca0088453b77d2e885dd05cf55dee983b69
Signed-off-by: Qi Zhou <qizhou@opensource.cirrus.com>