ASoC: cs35l45: Do not clear MBOX4 when handling MBOX IRQs
It is safer to have the MBOX as a one-way. If both sides writing
the to the MBOX4 for DSP event logging, especially, when that
happens quickly, there could be a situation where both the driver
and FW are trying to write a value concurrently.

The FW sends incrementing counts to MBOX4 when raising the MBOX IRQs,
the driver should receive a different value each time, therefore,
no need for the driver to clear the MBOX4 when dealing with the IRQs.
If the driver receives the same value, it indicates that the another
MBOX register is written to which should be handled in a separate
handler.

Change-Id: I3956bc5e2b128ff566db48003faf0ddc6632ea7f
Signed-off-by: Qi Zhou <qizhou@opensource.cirrus.com>