ASoC: cs40l25: Move handling of refclk errata into the ASoC code
When the clock is switched to running from MCLK an errata register write
needs to be applied based on the clock frequency. This was previously
contained in the main driver and predicated of the asp-bclk-freq-hz
devicetree property. Now the ASoC driver is responsible for the bclk
rate directly, the code to handle the errata should be moved there.
Otherwise the incorrect values will be written out if the BCLK rate
doesn't match that specified through devicetree.

Change-Id: I59ec935d4d29790a3379ac7b2f3efa4bb0e004b3
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>